+++ light-bootstrap-dashboard-react-master/src/routes.js
@@ -17,7 +17,6 @@
 */
 import Dashboard from "views/Dashboard.js";
 import UserProfile from "views/UserProfile.js";
-import TableList from "views/TableList.js";
 import Typography from "views/Typography.js";
 import Icons from "views/Icons.js";
 import Maps from "views/Maps.js";
@@ -25,6 +24,7 @@ import Notifications from "views/Notifications.js";
 import Upgrade from "views/Upgrade.js";
 import Routine from "views/Routine.js";
 import Report from "views/Report";
+import RoutineCreate from "views/RoutineCreate.js";
 const dashboardRoutes = [
   {
     upgrade: true,
@@ -32,56 +32,49 @@ const dashboardRoutes = [
     name: "Upgrade to PRO",
     icon: "nc-icon nc-alien-33",
     component: Upgrade,
-    layout: "/admin"
+    layout: "/admin",
   },
   {
     path: "/dashboard",
     name: "Dashboard",
     icon: "nc-icon nc-chart-pie-35",
     component: Dashboard,
-    layout: "/admin"
+    layout: "/admin",
   },
   {
     path: "/user",
     name: "User Profile",
     icon: "nc-icon nc-circle-09",
     component: UserProfile,
-    layout: "/admin"
-  },
-  {
-    path: "/table",
-    name: "Table List",
-    icon: "nc-icon nc-notes",
-    component: TableList,
-    layout: "/admin"
+    layout: "/admin",
   },
   {
     path: "/typography",
     name: "Typography",
     icon: "nc-icon nc-paper-2",
     component: Typography,
-    layout: "/admin"
+    layout: "/admin",
   },
   {
     path: "/icons",
     name: "Icons",
     icon: "nc-icon nc-atom",
     component: Icons,
-    layout: "/admin"
+    layout: "/admin",
   },
   {
     path: "/maps",
     name: "Maps",
     icon: "nc-icon nc-pin-3",
     component: Maps,
-    layout: "/admin"
+    layout: "/admin",
   },
   {
     path: "/notifications",
     name: "Notifications",
     icon: "nc-icon nc-bell-55",
     component: Notifications,
-    layout: "/admin"
+    layout: "/admin",
   },
   // {
   //   path: "/alert",
@@ -95,15 +88,22 @@ const dashboardRoutes = [
     name: "Routine",
     icon: "nc-icon nc-bell-55",
     component: Routine,
-    layout: "/admin"
+    layout: "/admin",
+  },
+  {
+    path: "/routinecreate",
+    name: "RoutineCreate",
+    icon: "nc-icon nc-bell-55",
+    component: RoutineCreate,
+    layout: "/admin",
   },
   {
     path: "/report",
     name: "Report",
     icon: "nc-icon nc-bell-55",
     component: Report,
-    layout: "/admin"
-  }
+    layout: "/admin",
+  },
 ];
 
 export default dashboardRoutes;
